coinview/requirements.txt
python-binance
binance-futures
TA-Lib
numpy
backtrader
flask
python-socketio==4.6.0
#python-engineio==3.13.2
flask-socketio==4.3.1
unicorn-binance-websocket-api